+++ nodeconductor/events/apps.py
@@ -1,8 +1,0 @@
-from __future__ import unicode_literals
-
-from django.apps import AppConfig
-
-
-class EventsConfig(AppConfig):
-    name = 'nodeconductor.events'
-    verbose_name = 'NodeConductor Events'