+++ nodeconductor/logging/elasticsearch_client.py
@@ -5,7 +5,7 @@ import logging
 from django.conf import settings
 from elasticsearch import Elasticsearch
 
-from nodeconductor.events.log import event_logger
+from nodeconductor.logging.log import event_logger
 
 
 logger = logging.getLogger(__name__)
@@ -33,7 +33,7 @@ class ElasticsearchResultList(object):
     def _get_client(self):
         if settings.NODECONDUCTOR.get('ELASTICSEARCH_DUMMY', False):
             # to avoid circular dependencies
-            from nodeconductor.events.elasticsearch_dummy_client import ElasticsearchDummyClient
+            from nodeconductor.logging.elasticsearch_dummy_client import ElasticsearchDummyClient
             logger.warn(
                 'Dummy client for elasticsearch is used, set ELASTICSEARCH_DUMMY to False to disable dummy client')
             return ElasticsearchDummyClient()
@@ -126,34 +126,6 @@ class ElasticsearchClient(object):
             verify_certs=elasticsearch_settings.get('verify_certs', False),
         )
 
-    def _get_permitted_objects_uuids(self, user):
-        """
-        Return list object available UUIDs for user
-        """
-
-        permitted_objects_uuids = event_logger.get_permitted_objects_uuids(user)
-
-        # XXX: this method has to be refactored, because it adds dependencies from iaas and structure apps
-        from nodeconductor.structure import models as structure_models
-        from nodeconductor.structure.filters import filter_queryset_for_user
-
-        if user.is_staff:
-            cusomter_queryset = structure_models.Customer.objects.all()
-        else:
-            cusomter_queryset = structure_models.Customer.objects.filter(
-                roles__permission_group__user=user, roles__role_type=structure_models.CustomerRole.OWNER)
-
-        permitted_objects_uuids.update({
-            'project_uuid': filter_queryset_for_user(
-                structure_models.Project.objects.all(), user).values_list('uuid', flat=True),
-            'project_group_uuid': filter_queryset_for_user(
-                structure_models.ProjectGroup.objects.all(), user).values_list('uuid', flat=True),
-            'customer_uuid': filter_queryset_for_user(
-                cusomter_queryset, user).values_list('uuid', flat=True),
-        })
-
-        return permitted_objects_uuids
-
     def _escape_elasticsearch_field_value(self, field_value):
         """
         Remove double quotes from field value
@@ -172,7 +144,7 @@ class ElasticsearchClient(object):
         return '%s:("%s")' % (field_name, '", "'.join(excaped_field_values))
 
     def _get_search_body(self, user, event_types=None, search_text=None, search_params=[]):
-        permitted_objects_uuids = self._get_permitted_objects_uuids(user)
+        permitted_objects_uuids = event_logger.get_permitted_objects_uuids(user)
         # Create query for user-related events
         query = ' OR '.join([
             self._format_to_elasticsearch_field_filter(item, uuids)