+++ nodeconductor/logging/perms.py
@@ -1,0 +1,6 @@
+from nodeconductor.core.permissions import StaffPermissionLogic
+
+
+PERMISSION_LOGICS = (
+    ('logging.Alert',  StaffPermissionLogic(any_permission=True)),
+)