+++ nodeconductor/quotas/models.py
@@ -4,12 +4,13 @@ from django.db import models, transaction
 from django.db.models import Sum
 from django.utils.encoding import python_2_unicode_compatible
 
+from nodeconductor.logging.log import LoggableMixin
 from nodeconductor.quotas import exceptions, managers
 from nodeconductor.core.models import UuidMixin, NameMixin
 
 
 @python_2_unicode_compatible
-class Quota(UuidMixin, NameMixin, models.Model):
+class Quota(UuidMixin, NameMixin, LoggableMixin, models.Model):
     """
     Abstract quota for any resource
 
@@ -49,6 +50,9 @@ class Quota(UuidMixin, NameMixin, models.Model):
 
     def __str__(self):
         return '%s quota for %s' % (self.name, self.scope)
+
+    def get_log_fields(self):
+        return ('uuid', 'name', 'limit', 'usage', 'scope')
 
 
 class QuotaModelMixin(models.Model):