+++ nodeconductor/server/base_settings.py
@@ -24,7 +24,7 @@ INSTALLED_APPS = (
     'django.contrib.staticfiles',
 
     'nodeconductor.landing',
-    'nodeconductor.events',
+    'nodeconductor.logging',
     'nodeconductor.core',
     'nodeconductor.backup',
     'nodeconductor.template',
@@ -32,6 +32,7 @@ INSTALLED_APPS = (
     'nodeconductor.quotas',
     'nodeconductor.structure',
     'nodeconductor.billing',
+    'nodeconductor.oracle',
     'nodeconductor.iaas',
     'nodeconductor.ldapsync',
 
@@ -52,7 +53,7 @@ MIDDLEWARE_CLASSES = (
     'django.middleware.common.CommonMiddleware',
     'django.middleware.csrf.CsrfViewMiddleware',
     'django.contrib.auth.middleware.AuthenticationMiddleware',
-    'nodeconductor.events.middleware.CaptureUserMiddleware',
+    'nodeconductor.logging.middleware.CaptureUserMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 )
@@ -209,6 +210,7 @@ CELERY_TASK_THROTTLING = {
 }
 
 NODECONDUCTOR = {
+    'EXTENSIONS_AUTOREGISTER': True,
     'DEFAULT_SECURITY_GROUPS': (
         {
             'name': 'ssh',